refactor: use lowercase name for App component file
